fix a bug where image transitions where totally warped in almost random cases. every case was IE.
this bug appears to be related to the speed of the download, and not the browser - it could very well happen in Firefox as well.

the problem is that the transition cycle was started when the DOM was ready - ie; when the browser had finished downloading the HTML of the page.

as the images were still downloading, the transition could attempt to show the second image before it has finished downloading. In IE, this resulted in a misshapen image which was too large.

I've changed the code so it runs when the window has completely finished loading, including all images.